Validations

+++ src/main/java/com/santanna/olympicgames/service/AthleteService.java
@@ -6,10 +6,12 @@ import com.santanna.olympicgames.domain.dto.UpdateAthleteDTO;
 import com.santanna.olympicgames.domain.entity.Athlete;
 import com.santanna.olympicgames.exceptions.ValidationException;
 import com.santanna.olympicgames.repository.AthleteRepository;
+import jakarta.validation.ConstraintViolationException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 @Service
 public class AthleteService {
@@ -21,28 +23,41 @@ public class AthleteService {
         return athleteRepository.findAll(page).map(AthleteRequestDTO::new);
     }
 
-    public AthleteRequestDTO getAthleteById(Long id){
+    public AthleteRequestDTO getAthleteById(Long id) {
         var athlete = athleteRepository.findById(id)
-                .orElseThrow(()-> new RuntimeException("Athlete not found"));
+                .orElseThrow(() -> new ValidationException("Athlete not found"));
         return new AthleteRequestDTO(athlete);
     }
 
-    public AthleteResponseDTO createAthlete(AthleteRequestDTO athleteDTO){
-        var athlete = new Athlete(athleteDTO);
-        athleteRepository.save(athlete);
-        return new AthleteResponseDTO(athlete);
+    public AthleteResponseDTO createAthlete(AthleteRequestDTO athleteDTO) {
+        try {
+            var athlete = new Athlete(athleteDTO);
+            athleteRepository.save(athlete);
+            return new AthleteResponseDTO(athlete);
+        } catch (ConstraintViolationException ex) {
+            throw new ValidationException("Error creating athlete: Invalid fields" + ex.getMessage());
+        }
     }
 
-    public AthleteResponseDTO updateAthlete(UpdateAthleteDTO athleteDTO){
-        var athlete = athleteRepository.findById(athleteDTO.id())
-                .orElseThrow(()-> new ValidationException("Athlete not found"));
-        athlete.updateAthleteData(athleteDTO);
-        athleteRepository.save(athlete);
-        return new AthleteResponseDTO(athlete);
+    public AthleteResponseDTO updateAthlete(UpdateAthleteDTO athleteDTO) {
+        try {
+            var athlete = athleteRepository.findById(athleteDTO.id())
+                    .orElseThrow(() -> new ValidationException("Athlete not found"));
+            athlete.updateAthleteData(athleteDTO);
+            athleteRepository.save(athlete);
+            return new AthleteResponseDTO(athlete);
+        } catch (ConstraintViolationException ex) {
+            throw new ValidationException("Error update athlete: Invalid fields" + ex.getMessage());
+        }
     }
 
-    public void deleteAthlete(Long id){
-        athleteRepository.deleteById(id);
-    }
+    public void deleteAthlete(Long id) {
+        boolean athleteExist = athleteRepository.findById(id).isPresent();
+        if (athleteExist) {
+            athleteRepository.deleteById(id);
+        } else {
+            throw new ValidationException("Athlete not found");
+        }
 
+    }
 }
